+++ index.php
@@ -38,17 +38,14 @@ endif; ?>">
 
 			<?php /* Start the Loop */ ?>
 			<?php while ( have_posts() ) : the_post(); ?>
-							<?php if ( ! is_sticky() ) : ?>
-				<?php
-
+				<?php if ( loose_show_sticky() ) :
 					/*
 					 * Include the Post-Format-specific template for the content.
 					 * If you want to override this in a child theme, then include a file
 					 * called content-___.php (where ___ is the Post Format name) and that will be used instead.
 					 */
 					get_template_part( 'template-parts/content-home', get_theme_mod( 'home_page_layout', 'masonry' ) );
-				?>
-							<?php endif; ?>
+				 endif; ?>
 			<?php endwhile; ?>
 
 			<?php the_posts_navigation(); ?>