+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log - Loose
 
+## [1.1] - 2017-06-28
+
+### Added
+- Option to not show sticky posts only in slider
+
+### Fixed
+- Logo size issue in menu and footer
+- Messed categories in slider
+
+## [1.0.4] - 2017-04-19
+
+### Changed
+- TGMPA version.
+
 ## [1.0.3] - 2017-04-09
 
 ### Added